Added settings for publishing to bintray
Basic support for publishing to bintray that can handle publishing to jcenter, maven-central and snapshots to oss.jfrog.org.

Bintray is configured in dryRun mode and that needs to changed before being able to publish.

User and Key needs to be passed to the gradle script via an environment variable or as a gradle property:
http://mrhaki.blogspot.se/2016/05/gradle-goodness-get-property-or-default.html

Note that versioning is not covered by this commit.